move radio back to radio-button-group
