+++ examples/radio-button-group/base.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import RadioButtonGroup from '~/components/radio-button-group'; // `~` is replaced with design-system-react at runtime
-import Radio from '~/components/forms/radio'; // `~` is replaced with design-system-react at runtime
+import Radio from '~/components/radio-button-group/radio'; // `~` is replaced with design-system-react at runtime
 
 class Example extends React.Component {
