+++ package.json
@@ -227,12 +227,6 @@
       "SLDS-component-path": "/components/forms/#flavor-checkbox"
     },
     {
-      "component": "forms/radio",
-      "status": "prod",
-      "display-name": "Radios",
-      "SLDS-component-path": "/components/forms/#flavor-radio"
-    },
-    {
       "component": "data-table",
       "status": "prod",
       "display-name": "Data Tables",
@@ -428,7 +422,12 @@
       "component": "radio-button-group",
       "status": "prod",
       "display-name": "Radio Button Groups",
-      "SLDS-component-path": "/components/radio-button-group"
+      "SLDS-component-path": "/components/radio-button-group",
+      "dependencies": [
+        {
+          "component": "radio"
+        }
+      ]
     },
     {
       "component": "tooltip",
